Fix for light entities and HACS updates #16

+++ custom_components/gtfs2/gtfs_helper.py
@@ -542,10 +542,16 @@ def get_datasources(hass, path) -> dict[str]:
 def remove_datasource(hass, path, filename):
     gtfs_dir = hass.config.path(path)
     _LOGGER.info(f"Removing datasource: {os.path.join(gtfs_dir, filename)}.*")
-    os.remove(os.path.join(gtfs_dir, filename + ".sqlite"))
-    os.remove(os.path.join(gtfs_dir, filename + "_temp.zip"))
-    os.remove(os.path.join(gtfs_dir, filename + "_temp_out.zip"))
-    os.remove(os.path.join(gtfs_dir, filename + ".sqlite"))
+    if os.path.exists(os.path.join(gtfs_dir, filename + ".sqlite")):
+        os.remove(os.path.join(gtfs_dir, filename + ".sqlite"))
+    if os.path.exists(os.path.join(gtfs_dir, filename + "_temp.zip")):     
+        os.remove(os.path.join(gtfs_dir, filename + "_temp.zip"))
+    if os.path.exists(os.path.join(gtfs_dir, filename + "_temp_out.zip")):        
+        os.remove(os.path.join(gtfs_dir, filename + "_temp_out.zip"))
+    if os.path.exists(os.path.join(gtfs_dir, filename + ".sqlite-journal")):        
+        os.remove(os.path.join(gtfs_dir, filename + ".sqlite-journal"))
+    if os.path.exists(os.path.join(gtfs_dir, filename + ".zip")):        
+        os.remove(os.path.join(gtfs_dir, filename + ".zip"))        
     return "removed"
     
 def check_extracting(hass, gtfs_dir,file):
